+++ spec/support/adapter_helper.rb
@@ -17,14 +17,19 @@ ensure
   Thread.kill(t)
 end
 
-shared_examples_for 'an adapter that call properly listener#on_change' do
+shared_examples_for 'an adapter that call properly listener#on_change' do |*args|
+  options = args.first || {}
   let(:listener) { mock(Listen::Listener) }
 
   context 'single file operations' do
     context 'when a file is created' do
       it 'detects the added file' do
         fixtures do |path|
-          listener.should_receive(:on_change).with([path])
+          if options[:recursive]
+            listener.should_receive(:on_change).once.with([path], :recursive => true)
+          else
+            listener.should_receive(:on_change).once.with([path])
+          end
 
           watch(listener, path) do
             touch 'new_file.rb'
@@ -35,13 +40,17 @@ shared_examples_for 'an adapter that call properly listener#on_change' do
       context 'given a new created directory' do
         it 'detects the added file' do
           fixtures do |path|
-            listener.should_receive(:on_change).with do |array|
-              array.should =~ [path, "#{path}/a_directory"]
+            if options[:recursive]
+              listener.should_receive(:on_change).once.with([path], :recursive => true)
+            else
+              listener.should_receive(:on_change).once.with do |array|
+                array.should =~ [path, "#{path}/a_directory"]
+              end
             end
 
             watch(listener, path) do
               mkdir 'a_directory'
-            # Needed for INotify, because of :recursive rb-inotify custom flag?
+              # Needed for INotify, because of :recursive rb-inotify custom flag?
               sleep 0.05 if @adapter.is_a?(Listen::Adapters::Linux)
               touch 'a_directory/new_file.rb'
             end
@@ -52,7 +61,11 @@ shared_examples_for 'an adapter that call properly listener#on_change' do
       context 'given an existing directory' do
         it 'detects the added file' do
           fixtures do |path|
-            listener.should_receive(:on_change).with(["#{path}/a_directory"])
+            if options[:recursive]
+              listener.should_receive(:on_change).once.with([path], :recursive => true)
+            else
+              listener.should_receive(:on_change).once.with(["#{path}/a_directory"])
+            end
 
             mkdir 'a_directory'
 
@@ -67,7 +80,11 @@ shared_examples_for 'an adapter that call properly listener#on_change' do
     context 'when a file is modified' do
       it 'detects the modified file' do
         fixtures do |path|
-          listener.should_receive(:on_change).with([path])
+          if options[:recursive]
+            listener.should_receive(:on_change).once.with([path], :recursive => true)
+          else
+            listener.should_receive(:on_change).once.with([path])
+          end
 
           touch 'existing_file.txt'
 
@@ -80,7 +97,11 @@ shared_examples_for 'an adapter that call properly listener#on_change' do
       context 'given a hidden file' do
         it 'detects the modified file' do
           fixtures do |path|
-            listener.should_receive(:on_change).with([path])
+            if options[:recursive]
+              listener.should_receive(:on_change).once.with([path], :recursive => true)
+            else
+              listener.should_receive(:on_change).once.with([path])
+            end
 
             touch '.hidden'
 
@@ -91,15 +112,21 @@ shared_examples_for 'an adapter that call properly listener#on_change' do
         end
       end
 
-      context 'given a file mode change' do
-        it 'does not detect the mode change' do
-          fixtures do |path|
-            listener.should_receive(:on_change).with([path])
+      unless options[:adapter] == :windows
+        context 'given a file mode change' do
+          it 'does not detect the mode change' do
+            fixtures do |path|
+              if options[:recursive]
+                listener.should_receive(:on_change).once.with([path], :recursive => true)
+              else
+                listener.should_receive(:on_change).once.with([path])
+              end
 
-            touch 'run.rb'
+              touch 'run.rb'
 
-            watch(listener, path) do
-              chmod 0777, 'run.rb'
+              watch(listener, path) do
+                chmod 0777, 'run.rb'
+              end
             end
           end
         end
@@ -108,7 +135,11 @@ shared_examples_for 'an adapter that call properly listener#on_change' do
       context 'given an existing directory' do
         it 'detects the modified file' do
           fixtures do |path|
-            listener.should_receive(:on_change).with(["#{path}/a_directory"])
+            if options[:recursive]
+              listener.should_receive(:on_change).once.with([path], :recursive => true)
+            else
+              listener.should_receive(:on_change).once.with(["#{path}/a_directory"])
+            end
 
             mkdir 'a_directory'
             touch 'a_directory/existing_file.txt'
@@ -124,7 +155,11 @@ shared_examples_for 'an adapter that call properly listener#on_change' do
     context 'when a file is moved' do
       it 'detects the file move' do
         fixtures do |path|
-          listener.should_receive(:on_change).with([path])
+          if options[:recursive]
+            listener.should_receive(:on_change).once.with([path], :recursive => true)
+          else
+            listener.should_receive(:on_change).once.with([path])
+          end
 
           touch 'move_me.txt'
 
@@ -137,8 +172,12 @@ shared_examples_for 'an adapter that call properly listener#on_change' do
       context 'given an existing directory' do
         it 'detects the file move into the directory' do
           fixtures do |path|
-            listener.should_receive(:on_change).with do |array|
-              array.should =~ [path, "#{path}/a_directory"]
+            if options[:recursive]
+              listener.should_receive(:on_change).once.with([path], :recursive => true)
+            else
+              listener.should_receive(:on_change).once.with do |array|
+                array.should =~ [path, "#{path}/a_directory"]
+              end
             end
 
             mkdir 'a_directory'
@@ -152,8 +191,12 @@ shared_examples_for 'an adapter that call properly listener#on_change' do
 
         it 'detects a file move out of the directory' do
           fixtures do |path|
-            listener.should_receive(:on_change).with do |array|
-              array.should =~ [path, "#{path}/a_directory"]
+            if options[:recursive]
+              listener.should_receive(:on_change).once.with([path], :recursive => true)
+            else
+              listener.should_receive(:on_change).once.with do |array|
+                array.should =~ [path, "#{path}/a_directory"]
+              end
             end
 
             mkdir 'a_directory'
@@ -167,8 +210,12 @@ shared_examples_for 'an adapter that call properly listener#on_change' do
 
         it 'detects a file move between two directories' do
           fixtures do |path|
-            listener.should_receive(:on_change).with do |array|
-              array.should =~ ["#{path}/from_directory", "#{path}/to_directory"]
+            if options[:recursive]
+              listener.should_receive(:on_change).once.with([path], :recursive => true)
+            else
+              listener.should_receive(:on_change).once.with do |array|
+                array.should =~ ["#{path}/from_directory", "#{path}/to_directory"]
+              end
             end
 
             mkdir 'from_directory'
@@ -186,7 +233,11 @@ shared_examples_for 'an adapter that call properly listener#on_change' do
     context 'when a file is deleted' do
       it 'detects the file removal' do
         fixtures do |path|
-          listener.should_receive(:on_change).with([path])
+          if options[:recursive]
+            listener.should_receive(:on_change).once.with([path], :recursive => true)
+          else
+            listener.should_receive(:on_change).once.with([path])
+          end
 
           touch 'unnecessary.txt'
 
@@ -199,7 +250,11 @@ shared_examples_for 'an adapter that call properly listener#on_change' do
       context 'given an existing directory' do
         it 'detects the file removal' do
           fixtures do |path|
-            listener.should_receive(:on_change).with(["#{path}/a_directory"])
+            if options[:recursive]
+              listener.should_receive(:on_change).once.with([path], :recursive => true)
+            else
+              listener.should_receive(:on_change).once.with(["#{path}/a_directory"])
+            end
 
             mkdir 'a_directory'
             touch 'a_directory/do_not_use.rb'
@@ -216,8 +271,12 @@ shared_examples_for 'an adapter that call properly listener#on_change' do
   context 'multiple file operations' do
     it 'detects the added files' do
       fixtures do |path|
-        listener.should_receive(:on_change).with do |array|
-          array.should =~ [path, "#{path}/a_directory"]
+        if options[:recursive]
+          listener.should_receive(:on_change).once.with([path], :recursive => true)
+        else
+          listener.should_receive(:on_change).once.with do |array|
+            array.should =~ [path, "#{path}/a_directory"]
+          end
         end
 
         watch(listener, path) do
@@ -234,8 +293,12 @@ shared_examples_for 'an adapter that call properly listener#on_change' do
 
     it 'detects the modified files' do
       fixtures do |path|
-        listener.should_receive(:on_change).with do |array|
-          array.should =~ [path, "#{path}/a_directory"]
+        if options[:recursive]
+          listener.should_receive(:on_change).once.with([path], :recursive => true)
+        else
+          listener.should_receive(:on_change).once.with do |array|
+            array.should =~ [path, "#{path}/a_directory"]
+          end
         end
 
         touch 'a_file.rb'
@@ -253,8 +316,12 @@ shared_examples_for 'an adapter that call properly listener#on_change' do
 
     it 'detects the removed files' do
       fixtures do |path|
-        listener.should_receive(:on_change).with do |array|
-          array.should =~ [path, "#{path}/a_directory"]
+        if options[:recursive]
+          listener.should_receive(:on_change).once.with([path], :recursive => true)
+        else
+          listener.should_receive(:on_change).once.with do |array|
+            array.should =~ [path, "#{path}/a_directory"]
+          end
         end
 
         touch 'a_file.rb'
@@ -274,7 +341,11 @@ shared_examples_for 'an adapter that call properly listener#on_change' do
   context 'single directory operations' do
     it 'detects a moved directory' do
       fixtures do |path|
-        listener.should_receive(:on_change).with([path])
+        if options[:recursive]
+          listener.should_receive(:on_change).once.with([path], :recursive => true)
+        else
+          listener.should_receive(:on_change).once.with([path])
+        end
 
         mkdir 'a_directory'
         touch 'a_directory/a_file.rb'
@@ -288,8 +359,12 @@ shared_examples_for 'an adapter that call properly listener#on_change' do
 
     it 'detects a removed directory' do
       fixtures do |path|
-        listener.should_receive(:on_change).with do |array|
-          array.should =~ [path, "#{path}/a_directory"]
+        if options[:recursive]
+          listener.should_receive(:on_change).once.with([path], :recursive => true)
+        else
+          listener.should_receive(:on_change).once.with do |array|
+            array.should =~ [path, "#{path}/a_directory"]
+          end
         end
 
         mkdir 'a_directory'